Moved the converter class to a separate file
